Consolidate common code in BufferedGraphics and re-enable some tests
