+++ src/System.Drawing.Common/src/System/Drawing/BufferedGraphicsContext.Windows.cs
@@ -10,13 +10,8 @@ using System.Threading;
 
 namespace System.Drawing
 {
-    /// <summary>
-    /// The BufferedGraphicsContext class can be used to perform standard double buffer rendering techniques.
-    /// </summary>
-    public sealed class BufferedGraphicsContext : IDisposable
+    public sealed partial class BufferedGraphicsContext : IDisposable
     {
-        private Size _maximumBuffer;
-        private Size _bufferSize;
         private Size _virtualSize;
         private Point _targetLoc;
         private IntPtr _compatDC;
@@ -31,96 +26,9 @@ namespace System.Drawing
         private const int BufferBusyPainting = 1; // The graphics buffer is busy being created/painting.
         private const int BufferBusyDisposing = 2; // The graphics buffer is busy disposing.
 
-        private static TraceSwitch s_doubleBuffering;
-
 #if DEBUG
         private string _stackAtBusy;
 #endif
-
-        /// <summary>
-        /// Basic constructor.
-        /// </summary>
-        public BufferedGraphicsContext()
-        {
-            // By defualt, the size of our maxbuffer will be 3 x standard button size.
-            _maximumBuffer.Width = 75 * 3;
-            _maximumBuffer.Height = 32 * 3;
-
-            _bufferSize = Size.Empty;
-        }
-
-        ~BufferedGraphicsContext() => Dispose(false);
-
-        /// <summary>
-        /// Internal trace switch for debugging
-        /// </summary>
-        internal static TraceSwitch DoubleBuffering
-        {
-            get
-            {
-                if (s_doubleBuffering == null)
-                {
-                    s_doubleBuffering = new TraceSwitch("DoubleBuffering", "Output information about double buffering");
-                }
-
-                return s_doubleBuffering;
-            }
-        }
-
-        /// <summary>
-        /// Allows you to set the maximum width and height of the buffer that will be retained in memory.
-        /// You can allocate a buffer of any size, however any request for a buffer that would have a total
-        /// memory footprint larger that the maximum size will be allocated temporarily and then discarded 
-        /// with the BufferedGraphics is released.
-        /// </summary>
-        public Size MaximumBuffer
-        {
-            get => _maximumBuffer;
-            set
-            {
-                if (value.Width <= 0 || value.Height <= 0)
-                {
-                    throw new ArgumentException(SR.Format(SR.InvalidArgumentValue, nameof(MaximumBuffer), value), nameof(value));
-                }
-
-                // If we've been asked to decrease the size of the maximum buffer,
-                // then invalidate the older & larger buffer.
-                if (value.Width * value.Height < _maximumBuffer.Width * _maximumBuffer.Height)
-                {
-                    Invalidate();
-                }
-
-                _maximumBuffer = value;
-            }
-        }
-
-        /// <summary>
-        /// Returns a BufferedGraphics that is matched for the specified target Graphics object.
-        /// </summary>
-        public BufferedGraphics Allocate(Graphics targetGraphics, Rectangle targetRectangle)
-        {
-            if (ShouldUseTempManager(targetRectangle))
-            {
-                Debug.WriteLineIf(DoubleBuffering.TraceWarning, "Too big of buffer requested (" + targetRectangle.Width + " x " + targetRectangle.Height + ") ... allocating temp buffer manager");
-                return AllocBufferInTempManager(targetGraphics, IntPtr.Zero, targetRectangle);
-            }
-
-            return AllocBuffer(targetGraphics, IntPtr.Zero, targetRectangle);
-        }
-
-        /// <summary>
-        /// Returns a BufferedGraphics that is matched for the specified target HDC object.
-        /// </summary>
-        public BufferedGraphics Allocate(IntPtr targetDC, Rectangle targetRectangle)
-        {
-            if (ShouldUseTempManager(targetRectangle))
-            {
-                Debug.WriteLineIf(DoubleBuffering.TraceWarning, "Too big of buffer requested (" + targetRectangle.Width + " x " + targetRectangle.Height + ") ... allocating temp buffer manager");
-                return AllocBufferInTempManager(null, targetDC, targetRectangle);
-            }
-
-            return AllocBuffer(null, targetDC, targetRectangle);
-        }
 
         /// <summary>
         /// Returns a BufferedGraphics that is matched for the specified target HDC object.
@@ -179,33 +87,6 @@ namespace System.Drawing
 
             return _buffer;
         }
-
-        /// <summary>
-        /// Returns a BufferedGraphics that is matched for the specified target HDC object.
-        /// </summary>
-        [SuppressMessage("Microsoft.Reliability", "CA2000:DisposeObjectsBeforeLosingScope")]
-        private BufferedGraphics AllocBufferInTempManager(Graphics targetGraphics, IntPtr targetDC, Rectangle targetRectangle)
-        {
-            BufferedGraphicsContext tempContext = null;
-            BufferedGraphics tempBuffer = null;
-
-            try
-            {
-                tempContext = new BufferedGraphicsContext();
-                tempBuffer = tempContext.AllocBuffer(targetGraphics, targetDC, targetRectangle);
-                tempBuffer.DisposeContext = true;
-            }
-            finally
-            {
-                if (tempContext != null && (tempBuffer == null || (tempBuffer != null && !tempBuffer.DisposeContext)))
-                {
-                    tempContext.Dispose();
-                }
-            }
-
-            return tempBuffer;
-        }
-
 
         /// <summary>
         /// Fills in the fields of a BITMAPINFO so that we can create a bitmap
@@ -471,12 +352,6 @@ namespace System.Drawing
             return hbmRet;
         }
 
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
         /// <summary>
         /// Disposes the DC, but leaves the bitmap alone.
         /// </summary>
@@ -612,18 +487,6 @@ namespace System.Drawing
             }
 
             _busy = BufferFree;
-        }
-
-        /// <summary>
-        /// This routine allows us to control the point were we start using throw away
-        /// managers for painting. Since the buffer manager stays around (by default)
-        /// for the life of the app, we don't want to consume too much memory
-        /// in the buffer. However, re-allocating the buffer for small things (like
-        /// buttons, labels, etc) will hit us on runtime performance.
-        /// </summary>
-        private bool ShouldUseTempManager(Rectangle targetBounds)
-        {
-            return (targetBounds.Width * targetBounds.Height) > (MaximumBuffer.Width * MaximumBuffer.Height);
         }
     }
 }
